+++ src/app/app.module.ts
@@ -6,16 +6,18 @@ import * as custom from '../assets/i18n/custom-en.json';
 import * as la from '../assets/i18n/la.json';
 import { DemoModule } from './demo/demo.module';
 import { mergeTranslations, I18N_EN, SidebarModule } from 'pastanaga-angular';
-import { DocModule } from './doc/doc.module';
+import { PaDocModule } from './doc/doc.module';
 import { Marker, Normalizer, Resolver, TraversalModule } from 'angular-traversal';
 import { DocResolver } from './doc/doc.resolver';
 import { DocMarker } from './doc/doc.marker';
 import { CommonModule } from '@angular/common';
+import { PaDocPagesModule } from './doc/doc-pages.module';
 
 @NgModule({
     imports: [
         DemoModule,
-        DocModule,
+        PaDocPagesModule,
+        PaDocModule,
         TraversalModule,
         SidebarModule,
         CommonModule,